+++ src/type/signUp/authEmailType.ts
@@ -1,0 +1,4 @@
+export interface AuthEmailType {
+  tableName: string;
+  userEmail: string;
+}
