+++ src/type/signUp/postVerifyCodeType.ts
@@ -1,0 +1,5 @@
+export interface PostVerifyCodeType {
+  tableName: string;
+  userEmail: string;
+  verificationCode: string;
+}
